+++ static/src/components/__tests__/LoginPane.spec.ts
@@ -1,6 +1,6 @@
 import { mount } from "@cypress/vue";
 import UserInformationLabel from "@/components/UserInformationLabel.vue";
-import LoginPane from "../LoginPane.vue";
+import LoginPane from "../LoginView.vue";
 
 describe("LoginPane", () => {
   it("exists", () => {
